fix(pe1168): looks like node but missing interface
This commit addresses issue #136. When extending nodes that were defined
with Node inference, the plugin can at times fail to detect the Node
interface. For example, if the original node being extented was defined
using SDL, extending it without adding `interfaces: Node` will mean that
GraphQl doesn't know that the node being created extends the Node type.
This mean that our plugin could cause errors to be thrown at build time,
as the newly created type was expected to be a Node type but was not.
For more information see the Gatsby documentation here:
https://www.gatsbyjs.com/docs/reference/graphql-data-layer/schema-customization/#gatsby-type-builders

It says, "When defnining top level types, don't forget to pass
`interfaces: ['Node']`, which does the same for Type Builders as adding
`implements Node` to SDL-defined types."