feat!: change getURL config from function to string
BREAKING CHANGE: This release improves support for Gatsby v4. As a result, the configuration for this plugin will need to be updated. This should be in most cases just a simple update. Instructions for this change can be found in the README under ["Upgrading from v1.x to v2"](./README.md#upgrading-from-v1x-to-v2).